+++ client/src/Components/PanelAdmin/ButtonsBar/ButtonsBar.module.css
@@ -1,0 +1,39 @@
+.ctn {
+  margin-top: 7px;
+  background-color: var(--white);
+  padding: 10px;
+}
+.ctn label{
+  cursor: pointer;
+}
+
+.elementsCtn {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.btnCtn {
+  display: flex;
+  width: 25%;
+}
+
+.btnBar {
+  color: #f5f5f5;
+  text-decoration: none;
+  font-size: 1rem;
+  cursor:  pointer;
+  margin: 0 10px;
+  background: #0578ac;
+  border-radius: 25px;
+  padding: 2%;
+  transition: all 0.3s ease-in-out;
+  width: 45%;
+  text-align: center;
+}
+
+.btnBar:hover {
+  background: #0f1a28;
+  box-shadow: 1px 1px 0px #bdbdbd;
+  transition: all 0.3s ease-in-out;
+}
