+++ client/src/Components/PanelAdmin/Comments/Comments.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
+import ButtonsBar from '../ButtonsBar/ButtonsBar';
 import './Comments.module.css';
 
 function Comments() {
   return (
     <div>
+      <ButtonsBar />
       <h1>soy el Comments</h1>
     </div>
   );
