--- client/src/Components/PanelAdmin/Header/Header.module.css
@@ -1,8 +1,0 @@
-.ctn {
-  margin-top: 7px;
-  background-color: var(--white);
-  padding: 10px;
-}
-.ctn label{
-  cursor: pointer;
-}
