+++ client/src/Components/PanelAdmin/Posts/Posts.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
+import ButtonsBar from '../ButtonsBar/ButtonsBar';
 import './Posts.module.css';
 
 function Posts() {
   return (
     <div>
+      <ButtonsBar />
       <h1>soy el Posts</h1>
     </div>
   );
