+++ client/src/Components/PanelAdmin/Users/Users.jsx
@@ -1,9 +1,11 @@
 import React from 'react';
+import ButtonsBar from '../ButtonsBar/ButtonsBar';
 import './Users.module.css';
 
 function Users() {
   return (
     <div>
+      <ButtonsBar />
       <h1>soy el Users</h1>
     </div>
   );
